pyproject.toml
[tool.poetry]
name = "extract-pdf-pages"
version = "0.1.0"
description = ""
authors = ["Daniel Rosenwald <danielrosenwald@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "extract_pdf_pages"}]

[tool.poetry.dependencies]
python = "^3.10"
pypdf = "^3.17.0"
typer = {extras = ["all"], version = "^0.16.0"}
pytest = "^8.4.1"

[tool.poetry.scripts]
extract-pdf-pages = "extract_pdf_pages:app"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
